requirements_full.txt
pydicom==2.4.4
pynetdicom==2.0.2
ttkbootstrap
numpy
pillow
openpyxl
